refactor: rename dto
