--- src/main/java/com/distinction/barbenheimer/DTO/HallIdDTO.java
@@ -1,26 +1,0 @@
-package com.distinction.barbenheimer.DTO;
-
-import com.distinction.barbenheimer.model.Seat;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.OneToMany;
-import jakarta.validation.constraints.NotEmpty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-@Builder
-public class HallIdDTO {
-
-    private Long id; // hall id
-
-    private int number; // hall number
-
-}
