+++ src/main/java/com/distinction/barbenheimer/controller/DemoController.java
@@ -1,17 +1,11 @@
 package com.distinction.barbenheimer.controller;
 
-import com.distinction.barbenheimer.DTO.MovieDetailsDTO;
 import com.distinction.barbenheimer.event.TicketPurchaseCompleteEvent;
-import com.distinction.barbenheimer.model.Purchase;
 import com.distinction.barbenheimer.repository.PurchaseRepository;
-import com.distinction.barbenheimer.service.MovieService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
 
 @CrossOrigin(origins = "*", allowedHeaders = "*")
 @RestController
