+++ src/main/java/com/distinction/barbenheimer/controller/SeatController.java
@@ -1,26 +1,15 @@
 package com.distinction.barbenheimer.controller;
 
-import java.time.*;
 import java.util.*;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.distinction.barbenheimer.DTO.SeatSelectDTO;
-import com.distinction.barbenheimer.DTO.SeatStatusDetailDTO;
-import com.distinction.barbenheimer.exception.AlreadyExistsException;
 import com.distinction.barbenheimer.model.CustomerDetail;
-import com.distinction.barbenheimer.model.MovieScheduleTime;
 import com.distinction.barbenheimer.model.Seat;
-import com.distinction.barbenheimer.model.SeatStatus;
 import com.distinction.barbenheimer.service.SeatService;
 
 @RestController
