+++ src/main/java/com/distinction/barbenheimer/dto/HallDetailDTO.java
@@ -1,4 +1,4 @@
-package com.distinction.barbenheimer.DTO;
+package com.distinction.barbenheimer.dto;
 
 public class HallDetailDTO { // this DTO returns the layout of seats in the hall
