+++ src/main/java/com/distinction/barbenheimer/dto/MovieImageDetailDTO.java
@@ -1,0 +1,16 @@
+package com.distinction.barbenheimer.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Builder
+public class MovieImageDetailDTO {
+
+    private String imageUrl;
+
+}
