+++ src/main/java/com/distinction/barbenheimer/dto/MovieScheduleDetailDTO.java
@@ -1,9 +1,5 @@
-package com.distinction.barbenheimer.DTO;
+package com.distinction.barbenheimer.dto;
 
-import com.distinction.barbenheimer.model.Hall;
-import com.distinction.barbenheimer.model.Movie;
-import com.distinction.barbenheimer.model.SeatStatus;
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -11,7 +7,6 @@ import lombok.NoArgsConstructor;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
-import java.util.List;
 
 @AllArgsConstructor
 @NoArgsConstructor
