+++ src/main/java/com/distinction/barbenheimer/dto/MovieShortDTO.java
@@ -1,4 +1,4 @@
-package com.distinction.barbenheimer.DTO;
+package com.distinction.barbenheimer.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
