+++ src/main/java/com/distinction/barbenheimer/dto/PurchaseDTO.java
@@ -1,4 +1,4 @@
-package com.distinction.barbenheimer.DTO;
+package com.distinction.barbenheimer.dto;
 
 
 import lombok.AllArgsConstructor;
