+++ src/main/java/com/distinction/barbenheimer/dto/SeatDetailDTO.java
@@ -1,12 +1,9 @@
-package com.distinction.barbenheimer.DTO;
+package com.distinction.barbenheimer.dto;
 
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-
-import java.util.List;
 
 
 @AllArgsConstructor
