+++ src/main/java/com/distinction/barbenheimer/dto/SeatNumberDTO.java
@@ -1,17 +1,19 @@
-package com.distinction.barbenheimer.DTO;
+package com.distinction.barbenheimer.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-@Data
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
-public class OngoingPurchaseTokenDTO {
+@Data
+public class SeatNumberDTO {
 
-    private String token;
-
+    private String rowCharacter;
+    
+    private int columnNumber;
 
 }
