+++ src/main/java/com/distinction/barbenheimer/repository/MovieScheduleTimeRepository.java
@@ -1,14 +1,11 @@
 package com.distinction.barbenheimer.repository;
 
-import com.distinction.barbenheimer.DTO.HallIdDTO;
-import com.distinction.barbenheimer.model.Movie;
 import com.distinction.barbenheimer.model.MovieScheduleDate;
 import com.distinction.barbenheimer.model.MovieScheduleTime;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.time.LocalTime;
 import java.util.List;
 import java.util.Optional;
