+++ src/main/java/com/distinction/barbenheimer/service/MovieScheduleService.java
@@ -1,7 +1,7 @@
 package com.distinction.barbenheimer.service;
 
-import com.distinction.barbenheimer.DTO.MovieScheduleDetailDTO;
-import com.distinction.barbenheimer.DTO.MovieScheduleDateDetailDTO;
+import com.distinction.barbenheimer.dto.MovieScheduleDetailDTO;
+import com.distinction.barbenheimer.dto.MovieScheduleDateDetailDTO;
 import com.distinction.barbenheimer.model.Movie;
 
 import java.time.LocalDateTime;
