+++ src/main/java/com/distinction/barbenheimer/service/MovieServiceImpl.java
@@ -5,8 +5,8 @@ import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
-import com.distinction.barbenheimer.DTO.MovieShortDTO;
-import com.distinction.barbenheimer.DTO.MovieTitleDTO;
+import com.distinction.barbenheimer.dto.MovieShortDTO;
+import com.distinction.barbenheimer.dto.MovieTitleDTO;
 import com.distinction.barbenheimer.exception.ResourceNotFoundException;
 import com.distinction.barbenheimer.model.MovieImage;
 import com.distinction.barbenheimer.repository.MovieImageRepository;
@@ -16,7 +16,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import com.distinction.barbenheimer.DTO.MovieDetailsDTO;
+import com.distinction.barbenheimer.dto.MovieDetailsDTO;
 import com.distinction.barbenheimer.model.Movie;
 import com.distinction.barbenheimer.repository.MovieRepository;
 import org.springframework.stereotype.Service;
