+++ src/main/java/com/distinction/barbenheimer/service/PurchaseService.java
@@ -1,8 +1,8 @@
 package com.distinction.barbenheimer.service;
 
-import com.distinction.barbenheimer.DTO.PurchaseResponseDTO;
-import com.distinction.barbenheimer.DTO.PurchaseDTO;
-import com.distinction.barbenheimer.DTO.SeatStatusDetailDTO;
+import com.distinction.barbenheimer.dto.PurchaseResponseDTO;
+import com.distinction.barbenheimer.dto.PurchaseDTO;
+import com.distinction.barbenheimer.dto.SeatStatusDetailDTO;
 import com.stripe.exception.StripeException;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.http.ResponseEntity;
