+++ src/main/java/com/distinction/barbenheimer/service/SeatService.java
@@ -3,17 +3,9 @@ package com.distinction.barbenheimer.service;
 import java.time.*;
 import java.util.*;
 
-import org.springframework.cglib.core.Local;
-
-import com.distinction.barbenheimer.DTO.HallScheduleSeatDetailDTO;
-import com.distinction.barbenheimer.DTO.SeatDetailDTO;
-import com.distinction.barbenheimer.DTO.SeatSelectDTO;
-import com.distinction.barbenheimer.DTO.SeatStatusDetailDTO;
+import com.distinction.barbenheimer.dto.SeatDetailDTO;
 import com.distinction.barbenheimer.model.CustomerDetail;
-import com.distinction.barbenheimer.model.Hall;
-import com.distinction.barbenheimer.model.MovieScheduleTime;
 import com.distinction.barbenheimer.model.Seat;
-import com.distinction.barbenheimer.model.SeatStatus;
 
 public interface SeatService{
