+++ src/test/java/com/distinction/barbenheimer/application/MovieServiceTest.java
@@ -1,23 +1,14 @@
 package com.distinction.barbenheimer.application;
 
-import com.distinction.barbenheimer.DTO.MovieShortDTO;
-import com.distinction.barbenheimer.model.Movie;
 import com.distinction.barbenheimer.repository.MovieRepository;
 import com.distinction.barbenheimer.service.MovieServiceImpl;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.time.LocalDateTime;
-import java.time.Month;
-import java.util.ArrayList;
-import java.util.List;
-
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 public class MovieServiceTest {
